pk: Add prepared SQL statement cache
- Add simple cache of prepared SQL statements.  query_free() will reset
the query and insert it into the cache, and query() will try to allocate
from the cache first.  This should prevent commonly-used queries from
being repeatedly reparsed.

- Store a copy of the original SQL query string in struct query

- Add query_flush() function to empty the prepared statement cache.  This
must be called before sqlite3_close() if the database to be closed might
have prepared statements in the cache, or the close will fail.