Integrate distribution-specific openisr-config changes into mainline
- Add --with-kbuild-wrapper configure option, specifying the packaging
system to be used for building the kernel modules.  Only allow modules to be
built if --with-kbuild-wrapper=none (the default).  Using values other than
"none" requires special handling that the build system does not provide,
since this option is intended to support packaging scripts that can supply
this special handling.  So, for most users who might run the configure
script directly, this option's only effect will be to cause breakage.

- Integrate Debian and DKMS module-building methods into
conf/openisr-config.in, and choose the correct method depending on the
--with-kbuild-wrapper setting

- Have the redhat .spec file and the debian rules file use
--with-kbuild-wrapper

- Remove Debian- and Fedora-specific patches for openisr-config