locksrv/commit: Be safe w.r.t. clients with old nonces
- Add mandatory -N parameter specifying the nonce associated with the
parcel lock.  If the lock is not held or the nonce does not match, refuse
to do the commit.

- Commit from $parceldir/cache/$nonce rather than $parceldir/cache, so that
a client which (continues to) upload after its nonce has been revoked will
not interfere with uploading done by another client with the correct nonce

- After the commit succeeds, remove $parceldir/cache in order to flush
old partial uploads from clients with other nonces.  This should be safe,
since right after a commit, no one else can legitimately be uploading.

- Fail if the cachedir does not exist.  We used to silently succeed, but
this makes little sense and no one depended on this behavior.