pk: Split misleading print_progress() function into two alternatives
- print_progress() shows a progress meter which claims to represent the
number of MB to be processed, but the MB value is based on the number of
chunks and the incorrect assumption that all chunks are chunksize bytes.
We can't always give a real MB number because, as in hoard(), we may not
know how many bytes must be transferred.  Eliminate print_progress() and
provide print_progress_chunks() (prints progress in chunks, for when we
don't know the actual sizes) and print_progress_mb() (prints progress in
MB, for when we do).  Update callers.

- Have print_progress_chunks() update once a second rather than when
the output changes, since updating once per chunk is too often and it's
not clear how we should round off (e.g. by updating only every 100 chunks).
Have print_progress_mb() do the same for consistency.