pk: Remove double transaction in hoard_put_chunk()
hoard_put_chunk() writes to the hoard cache in a three-step process: first
it allocates the cache slot, then it writes to the hoard cache (outside of a
hoard transaction), then it updates the cache slot metadata.  This was done
so that the hoard index locks would not be held while the data write was
occurring.  However, this was probably premature optimization, since in most
cases the data write will just go into the page cache and return quickly --
and the three-step commit complicates the code substantially.  So, we remove
it:

- Update hoard_put_chunk() to perform the entire put operation in one
SQL transaction

- Have allocate_chunk_slot() just return the chosen offset, rather than
(redundantly) updating the chunks table to claim it.  Have hoard_put_chunk()
set the referenced flag while updating chunk metadata.

- Fold deallocate_chunk_offset() into its sole remaining caller

- Eliminate the special meaning of (hoard.chunks.referenced == 1 AND
hoard.chunks.tag ISNULL) as a cache slot which has been allocated and
is being written to.  Remove the dangling-slot cleanup code in
hoard_try_cleanup(), and update check_hoard() to always correct this case
(since we no longer need to wait until no other PK instances are running).
Update unused-slot test in allocate_chunk_offset().