pk: Fix hoard cache corruption caused by commit 24536b803
- Fix bug in hoard_put_chunk() introduced by 24536b803.  The effect of
this bug was:

1.  (Likely) The metadata for newly written chunks would not be written
to the cache index, and the slots would be reclaimed as orphan by
hoard_try_cleanup(), or

2.  (Less likely) The metadata for newly written chunks would be written to
the wrong place in the cache index, causing the index to diverge from the
hoard cache. The originally-allocated slots would be reclaimed as orphan.

Users who ran affected revisions should run "isr checkhoard -c" after
updating.