pk: Have the busy handler give up after 50 ms
- Update busy handler to fail the query after 50 ms.  SQLite can have ABBA
deadlocks between two processes which try to lock attached databases in
opposite orders, and the busy handling code in SQLite doesn't detect this
case.  The only apparent way to work around this from our code is to have a
finite backoff ourselves, after which we release all of our locks and try
again.

- Remove the busy handler highwater counter, since it no longer makes sense
with a finite timeout

- Add a counter of the number of queries which timed out in the busy handler