pk: Fix tight busy-loop in case of duelling write-after-read transactions
- The SQLite busy handler is called when a shared lock cannot be obtained on
a database, but is not called when a shared lock cannot be promoted to a
reserved lock.  Thus, when query_busy() returns true, we can't just retry
immediately: we need to have a delay first to allow the contention to clear.
Add query_retry() helper function which returns 1 iff query_busy() would,
but also waits a random amount of time before returning if we are indeed to
retry the query or transaction.  query_retry() should only be called once
the transaction's locks have been released (by freeing the query or rolling
back the transaction), because otherwise we'll be sleeping with locks held
(which won't help anything).  query_busy() is still available for testing
the error code without sleeping.

- Update most users of query_busy() to call query_retry().
_begin()/_commit()/_rollback still use query_busy().

- Rearrange copy_for_upload(), validate_keyring(), and validate_cachefile()
error paths to release locks before calling query_retry()

- Seed the libc random number generator from sql_init(), since we use it
in query_retry()