pk: Map SQLITE_IOERR_BLOCKED to SQLITE_BUSY
- Rename set_busy_handler() to sql_setup_conn(), and add code to enable
extended result codes on the database connection

- Have query_next() collapse SQLITE_IOERR_BLOCKED into SQLITE_BUSY, to
ensure that a retry will happen in this case.  IOERR_BLOCKED is issued in a
corner case in which SQLite cannot obtain an exclusive lock in order to
spill cache pages.  Extended result codes must be enabled (as above) to
distinguish this case from other SQLITE_IOERR cases.