pk: Fix race in hoard_put_chunk() leading to hoard cache corruption
- In hoard_put_chunk(), don't write the chunk data to the hoard cache until
after the metadata updates complete (but before they are committed).  This
fixes a race in which we obtain an offset to write to, finish the write, and
then fail with SQLITE_BUSY when we try to update metadata.  In this case, we
have written data to an offset which may already have been claimed (and
written) by another PK process, leading to hoard cache corruption. This fix
also improves performance in the presence of hoard_put_chunk() query
retries, since we don't do chunk I/O for every failed attempt.

- hoard_put_chunk(): Correct error message produced if the metadata update
fails