client: Print informative message if client does not have SSH keys
- Add CHECK_AUTH option to srun() which redirects SSH stderr and checks
it for "Permission denied" messages.  If any are found, it calls errexit()
with a helpful error message.

- Use CHECK_AUTH from need_remote_parcel(), which executes the first SSH
connection to be made by most subcommands

- Use CHECK_AUTH from listparcels(), which doesn't use need_remote_parcel()

- Add need_credentials() helper, which executes a dummy SSH connection for
the sole purpose of running CHECK_AUTH.  Call it at startup for "isr
passwd", which needs the check but doesn't normally use srun() at all

- Remove REDIRECT_STDERR option from srun(), since it is subsumed by
CHECK_AUTH for the single caller that used it

- Document need_remote_parcel()