client: Purge completion caches when appropriate
- At startup, remove all completion caches older than 30 minutes.  The
completion code does this too, but only when a cache is accessed.  If remote
parcel completion is not used on a client host for a long time, stale
completion caches could stay around, and the user may not want that
information to remain on-disk.

- Purge the completion cache for a server/user pair after authenticating
to it, to make sure negative results are flushed

- Purge all completion caches after "isr deauth", since none of those
parcels are now accessible