pk: Add generic signal-handling infrastructure
- Move nexus_init() signal-handler setup code to a utility function.  Rename
signal_handler() to nexus_signal_handler().

- Add generic signal handler which sets a state flag, and pending_signal()
helper which checks the flag and logs (only once) that PK has been
interrupted

- In main(), install generic signal handler for INT, TERM, and HUP (but
not QUIT) and block USR1 and USR2.  Check for pending signals after module
initialization, and shut down if any are present.

- Have nexus_init() check pending_signal() after overriding the generic
signal handlers, and return the new PK_INTERRUPT error code if a signal
is pending

- Update shutdown code to ensure that if we die on a signal, that fact
is propagated to our parent

- Update fork_and_wait() so that the parent propagates the fact that the
child died on a signal