pk: Handle caught signals by propagating them to SQLite
Add code to the SQL wrapper to convert caught signals into SQL errors, which
should lead to a rollback and a program abort.  This code is active in every
non-trivial mode except run mode, which uses its own signal handlers.  Once
a signal has been caught, we fire the SQLite interrupt mechanism on every
new or existing query, and also in long-running queries which are currently
in execution.

- Have query_next() check for pending signals.  If one is found, use
sqlite3_interrupt() to try to force SQLite to error out the query and cause
an automatic rollback.  This will often not work, depending on the query.
If it does work, the query will return SQLITE_INTERRUPT.  We'll keep trying
on every subsequent query, so if it doesn't work on this one, it may work
on the next.

- Add a SQLite progress handler which fires during long-running queries.
When it fires, it checks for pending signals; if there are any, we tell
SQLite to error out the query with SQLITE_INTERRUPT and do a rollback.

- Have pk_log_sqlerr() squash SQLITE_INTERRUPT errors

- Have rollback() squash SQL errors if the previous SQL error was
SQLITE_INTERRUPT, since in this case an automatic rollback has already
occurred

- Remove obsolete comment