client: Create parceldir and parcel.cfg with secure permissions
- In checkout(), create the parceldir mode 700.  (hoard() uses mktempdir(),
which already uses mode 700.) This is a belt-and-suspenders change in case
~/.isr is not mode 700.

- Add an argument to touch(), specifying the mode of the file to be created.
Update callers to explicitly pass the previous default (666).

- In checkout(), create the parcel.cfg mode 600.  This is likewise
belt-and-suspenders with regard to ~/.isr, but note that the parcel.cfg is
created in ~/.isr and then moved into $parceldir, so setting $parceldir
mode 700 is not sufficient.  On the other hand, hoard() writes the
parcel.cfg directly into the (mode 700) parceldir, so no such change is
needed there.