pk: Batch metadata updates for hoard cache chunk insertions
With the synchronous pragma set to NORMAL or FULL, insertion of a couple
hundred megs of data into the hoard cache results in thousands of fsync()
calls from SQLite.  If synchronous is set to OFF, these calls are not made,
but we run the risk of irreparable hoard cache index corruption if the
system encounters a hard lockup.  With the fsync() calls enabled, however,
the time spent waiting for them accounts for more than half of the total
runtime of copy_for_upload(), and probably about as much for hoard() etc.

To solve this problem, we introduce a "slot cache".  We restore the former
practice of writing to the hoard cache index in two stages: one stage
allocates the cache slot, and the other stage updates the metadata so that
the chunk is accessible when queried for.  However, the first stage now
allocates batches of 128 cache slots and stores them in the slot cache
(which is a temporary table, thus avoiding the fsync() overhead).
Insertions of individual chunks cause updates only to the slot cache, not to
the hoard cache index.  When the slot cache fills up or PK terminates, the
hoard index is updated to reflect the final disposition of each cached slot.
This change reduces the number of fsync() calls needed to a handful, while
still preventing index corruption and bounding the amount of data which will
be lost on a system crash.  (It should also reduce write-lock contention for
the hoard cache index.)

- Update hoard_try_cleanup() and check_hoard() for new cache slot allocation
semantics

- Create the slot cache table at startup, and flush the slot cache back to
the hoard cache index at shutdown

- Search both the slot cache and the hoard cache when checking for tag
presence in hoard_get_chunk() and hoard_put_chunk()

- Allocate chunk slots from the slot cache, flushing & refilling it if full

- When invalidating a cache chunk in hoard_get_chunk() error paths,
perform the correct invalidation actions depending on where the chunk
metadata came from

- When updating the access timestamp in hoard_get_chunk(), update the
correct table depending on the metadata source

- Add transaction to hoard_try_cleanup()