conf/completion: Don't assume subcommand is the first argument
- Add _isr_find_bareword() helper.  Eliminate _isr_offer_parcel_parameter()
and use _isr_find_bareword() instead.

- We can't assume that the subcommand is the first argument, because the
user might put options first.  So, assume there will be up to two barewords
on the command line: the subcommand and the parcel, in that order.  Offer a
subcommand if there are no barewords, or if we're editing the first one, or
if there's only one and the cursor is before it.  Offer a parcel if there is
only one bareword and the cursor is after it, or there are two and we're
editing the second one.

- Always offer to complete option parameters even if we don't know the
subcommand, since -q/-v always work