upgrade-0.9: Fix checkparcel warnings on generated parcel
- upgrade-0.9 could produce parcels whose hdk directories for non-last
versions were not completely sparse -- that is, they contained some chunks
which had not been modified in the subsequent parcel version.  This did
not compromise the integrity of the resulting parcel, but would cause it
to fail checkparcel validation; if it occurred in the second-to-last parcel
version, it would prevent the parcel from being checked in.  Fix this.