sqlite: Add fix for ticket #2371
- ATTACH returns SQLITE_ERROR rather than SQLITE_BUSY when the database to
be attached (or the main database) is EXCLUSIVE-locked (#2096, #2914).  This
is difficult to fix in SQLite, but we can catch this case by comparing
sqlite3_errmsg() to "database is locked".  However, sqlite3_errmsg() is not
correctly updated until sqlite3_reset() or sqlite3_finalize() is called,
contrary to the documented conventions for sqlite3_prepare_v2(), which means
that PK retrieves incorrect error messages in these (and probably other)
cases.  Fix SQLite to propagate the correct error message at sqlite3_step()
time.