pk: query_retry(): Make SQLITE_ERROR/"database is locked" retryable
- When executing an ATTACH, SQLite returns SQLITE_ERROR with errmsg set to
"database is locked" when it really means SQLITE_BUSY.  Catch this case in
query_retry().  (We can't catch it in attach() because we need the random
backoff from query_retry().  Plus, doing it in query_retry() should be more
robust if the same issue arises with other queries.)  This fixes random
failures occurring when trying to attach the hoard cache index while it is
EXCLUSIVE-locked.