pk: Don't create local cache in examine mode
- Make sure open_cachedir() only creates a missing local cache if the
current mode acquires the PK lock.  This should prevent races in which "isr
stat" creates the local cache at the same time as another PK process in
another mode.

- open_cachedir(): Don't complain about (exists(cache_image) XOR
exists(cache_index)) unless we have the PK lock, since this may occur due
to races with the process that does