pk: Add --splice option to validate mode
- Add --splice option to validate mode.  If this is used along with --full
and PK finds tag check failures, it will revert the failed chunks to their
state at the last checkout or sync.  (It does this by deleting the chunk
from the cache index and reverting its keyring entry, so that the next
access will cause a refetch.) This renders the disk inconsistent, but for
recovering data from a damaged local cache it may be useful to have stale
data rather than inaccessible data.  (Nexus returns I/O errors for chunks
which experience tag check failures.) The damaged flag is still set in this
case, so the user will not be able to check in the parcel. Currently
--splice has no effect if not used with --full.

- Do a commit() from validate_cachefile(), since we may indeed be updating
the database