pk: Prefix log messages with useful strings
- When writing to the log file, add a "TYPE:" field after the PID, where
"TYPE" is the uppercase name of the message's logtype (as understood by
parse_logtype()) and ":" is a colon.  This allows, for example, all error
messages to be easily grepped from the log file.  The uppercase makes the
logtype easier to read and to grep for, and the colon makes it easier to
separate the header information from the log message.

- When writing to stderr, prepend "PK: " to every log message, to make
it easier to separate the client's messages from PK's messages