README: Document unsafety of host suspend
- Add text explaining that it is not safe to suspend the host while parcels
are running.  (The process freezer used in the Linux suspend implementation
suspends user-space processes and then calls sys_sync().  If I/O is pending
against Nexus, this leads to a deadlock, since Parcelkeeper is already
frozen.  Only in very recent kernels is there even an interface that allows
us to intercept and cancel suspend before userspace is frozen.  We could
make PK non-freezable, but then we'd need to find a way to go back and
freeze it before memory image writeout starts, or else we risk corruption of
the suspend image.  In summary, there's no good way to solve this problem as
long as Nexus requires PK to be available in order to make progress with
writeback.)