Support chained scatterlists
Linux 2.6.24 introduced chained scatterlists where chains are identified
based on the lower bits in the page pointer. A set of new operations
were introduced to query (sg_page) and modify (sg_set_page) the page
pointer as well as initialize (sg_init_table) and  iterate correctly
(sg_next) over the chained lists.

The sg_next iterator is not needed in all places we walk the
scatterlists because the internally allocated scatterlists are never
chained. We only use sg_next in scatterlist_copy where we handle the
scatterlist from the original IO request.