pk: Ignore SIGTTOU and catch SIGHUP in run mode
- Add SIGTTOU to the list of ignored signals in run mode.  Usually,
the VMM runs in the same process group as PK.  However, if vmcommand
is set to /bin/sh in .openisrrc, the "VMM" puts itself in its own
process group, which means that PK is now a background process.  Then,
if PK tries to log a message to the console, its process group will
get SIGTTOU, causing PK to be suspended and further I/O to the Nexus
device to block.  (This will manifest as a spontaneous "Stopped"
message printed by the parent shell, followed by a shell prompt.
Running "fg" will resume PK's process group, but will also move it to
the foreground, which the "VMM" will not expect.)  However, by ignoring
SIGTTOU in PK, we tell the kernel that we want to be able to write
to the terminal even if we're in the background, and PK's process group
will not receive SIGTTOU.

- Catch SIGHUP in run mode instead of ignoring it.  This ensures that we
will shut down properly if our session leader (the parent shell) dies.