vmm/vmware: Auto-generate and -update VMX and VMDK files
- Autogenerate VMX/VMDK if there is no valid VMX file in CFGDIR

- Add $CFGDIR/uuid.vmware file, which tracks the UUID the CFGDIR thinks
it has.  If the parcel uuid no longer matches the CFGDIR UUID, the
parcel has been cloned; in this case, strip all unique identifiers from
the VMX and VMDK files so that VMware can regenerate them.  (We also
update the displayName.)

- Likewise strip unique identifiers for parcels without a uuid.vmware
file; this will cause *all* existing parcels to have their IDs
regenerated.  This is the cleanest way to make sure there are no
existing duplicates in the system.

- Add additional NAME and SECTORS inputs to the VMM driver in run mode
to support all of this

- Clean out VMware logs and core files from the cfg directory at runtime