libvdisk: Obtain wrapped symbols lazily
- Obtain pointers to wrapped functions just before first use, rather
than in the constructor.  This is necessary because libselinux has a
constructor which calls functions we have wrapped (when SELinux is
enabled), *before* our constructor has had a chance to run; this would
cause us to call through a NULL pointer and crash.

- Don't call exit() if we fail to obtain a symbol -- just warn() and
then crash.  Since we're calling dlsym() lazily, this can now occur
at any point in the program's execution, not just at initialization,
so we shouldn't make assumptions about particular exit codes.  In any
event, this should never happen.

- Move dlsym() function pointers to the individual functions that use
them, since they no longer need to be global