client: Only enable unsafe mode for appropriate subcommands
- Add $unsafe_opt global variable, which is the raw (tri-state) output
from the command-line parser.  $unsafe_mode is now simply a boolean which
specifies whether to enable unsafe mode.

- Move enabling of unsafe mode (including printing the requisite message)
to a function, and only call the function for subcommands which are
affected by it.  This prevents the warning message from being printed
when it is irrelevant to the operation being performed.

- Update usage message, man page, and bash completions to correctly
reflect the set of subcommands which care about unsafe mode