vmm/virtualbox: More aggressively ensure that the VM is unregistered
- No longer assume in unregister_machine() that the VM is in a known,
completely registered state.  Instead, query whether the VM and disk
image are registered, and only unregister them if they are.  This makes
it safe to call unregister_machine() as a cleanup measure.

- Call unregister_machine() early in run() to clean up from previous
crashes

- Load the XPCOM bindings and attempt to unregister the VM from cleanup().
Always discard the memory image during unregistration in this case,
because there's no way to safely update the suspended flag (since the
parcel might have been resumed with another VMM in the interval).

- Split savefile handling into its own function

- No longer pass the vmid and diskid around: the vmid is just vmm.UUID,
and the diskid can be looked up from disk_path