vmm: Improve SIGINT handling in run mode
- client: When executing the VMM driver in run mode, ignore SIGINT in
the client itself.  This causes the client to wait for the driver's
return status rather than bailing out immediately, so that the VMM
driver can react to SIGINT any way it wants.

- pk: Ignore SIGINT in run mode.  Without this, PK will get SIGINT at
the same time as the client and the VMM driver, and will start shutting
down.  However, the VMM driver may want to close and reopen the block
device during shutdown (vbox does this), and when it tries to reopen the
device it will already be gone.  The client will send SIGTERM to PK when
it completes its own shutdown, so we can ignore the SIGINT that kicks
everything off.  We still catch SIGQUIT in PK (and we have never caught
it in the client), so QUIT can still be used to force immediate shutdown.

- vmm.pm: Document that run_program() ignores SIGINT/SIGQUIT while the
child is running

- vmm.py: Ignore SIGINT/SIGQUIT in run_program() while the child is
running, since this is what the caller usually wants