vmm/virtualbox: Fixes for VirtualBox OSE
Fix several problems with the virtualbox driver and VirtualBox OSE.
Even with these fixes, the driver does not work properly with OSE
(OSE throws an exception when we attempt to allocate a session object).
But this gets us closer.

- Under VBox OSE, vbox.version contains a trailing "_OSE" in
contravention of the API documentation.  Handle this.

- Properly handle OSE's lack of USB support in create_machine()

- Properly handle a failure to set the network adapter type in
create_machine(), since the Intel adapter isn't supported by OSE