vmm/virtualbox: Properly resume VMs with inaccessible CD/floppy images
- If a VM XML file refers to an unregistered CD or floppy image,
vbox.openMachine() will fail.  Before opening the machine, walk through
the XML and unbind any CD or floppy image whose UUID is not recognized
by VirtualBox.

- VBox will refuse to resume a VM with an inaccessible CD or floppy
image.  At registration time, check all associated CD and floppy
images for accessibility, and unbind any image which is inaccessible.

- At registration time, check all associated host CD and floppy drives
to ensure that the device node exists and we have at least read
permission to it.  If not, unbind the drive from the VM.