vmm/virtualbox: Improve unregistration after a VBox upgrade
- Move the "cleanup unregistration" call as early as possible in run(),
since we shouldn't do XML fixups on a registered VM

- When unregistering a machine which was registered by a version of VBox
with an earlier settings version, it's more important to remove the
dangling registration than to ask the user whether to upgrade: the
registration can't be removed without upgrading the settings version,
and it needs to be removed.  In this case, warn the user that the upgrade
is occurring and save a backup copy of the settings file.