vmm/vmware: Run VMware in its own background process group by default
VMware displays an error if it gets SIGINT while it is shutting down in
response to a different SIGINT.  As a result, if the user types Ctrl-C
in the isr xterm, VMware displays an error message (before exiting
anyway), since it gets:

1.  One SIGINT from the terminal driver,
2.  One SIGINT forwarded from vmm/vmware, and
3.  One SIGINT forwarded from isr via vmm/vmware.

We fix this by ensuring that VMware will only get one SIGINT in this
case.

- Only forward SIGINT from vmm/vmware to VMware the first time we get it.
This prevents VMware from getting SIGINT #3.

- Run VMware in its own process group, with stdin/stdout/stderr sent to
/dev/null, unless -v is specified on the isr command line.  This prevents
VMware from getting SIGINT #1.  We need to redirect to /dev/null rather
than ignoring SIGTTIN/SIGTTOU, because something in the VMware stack
resets signal handlers for these signals, causing "isr resume -F" to
suspend the VMware processes (SIGTTOU) on VMware Player 2.5 (but not
2.0).  If VERBOSE is set, we run as before so that the user can see
VMware output; Ctrl-C will still produce an error message in this case.

- Print a message when forwarding SIGINT to VMware