dirtometer: Fix bitmap sizing
- Calculate the dimensions of the block bitmap based on the image's size
allocation rather than the size of the entire window

- Don't call update_img() at startup, since we need to wait for the
size-allocate event to know what to draw

- Only draw those pixels actually showing chunks; fill the remainder
of the last line with transparent pixels and don't fill subsequent lines
at all.  As a result, we may use less than our allocated size.

- Always render the entire bitmap into the image, even if it's larger
than our allocation

- Align the image to the top of its allocation so that it doesn't float
in the center if the allocation gets too big