vmm/virtualbox: Ask user to try again if session lasts < 20 seconds
- Sometimes VirtualBox crashes at startup without returning an error
to the user or to the virtualbox driver.  (An error is logged in
VBox.log, but not in a consistent format.)  These crashes usually
result from problems restoring the memory image, and cause the memory
image to be deleted.  To reduce user confusion in this case, notice
when a resume lasts less than 20 seconds and print a message suggesting
that the user retry resuming the parcel.