Work around udev bug causing device node owner not to be set
- udev versions 137-142, and thus Fedora 11 and Ubuntu 9.04, fail to
expand string substitutions on OWNER keys in udev rules.  As a result,
the existing mechanism for dynamically setting ownership of Nexus block
devices will not work on these systems.  Instead, have the udev rules
run a script which waits in the background for the device node to be
created and then corrects its ownership.