sql: Add query_interrupt()/query_clear_interrupt()
- Replace stray pending_signal() checks in sqlite/wrapper.c with
query_interrupt() and query_clear_interrupt() functions which set and
clear an interrupt flag on a particular database.  While the flag is
set, queries against that db will be interrupted at every possible point.
The set/clear functions are safe to call from other threads and from
signal handlers in the same thread.