sql: Fix "isr_admin makeparcel" SQL error
isr_admin makeparcel performs a compound query in which it creates a
table and then creates an index on that table.  This fails because
sql_split() parses every query in a compound query before any of them
are executed, and SQLite generates a parse failure on an attempt to
create an index on a nonexistent table.  Change compound-query handling
so that we parse each query as it is executed.

- sql: Rename sql_split() to sql_head(), and have it return the first
SQL statement plus a tail pointer

- sql: Convert sql_head() and query_parameter_count() to run inside a
transaction, since we'll be running these incrementally now

- sql: Eliminate g_list_to_strv() helper

- tools/query: Split query string and determine parameter counts
incrementally rather than up front