vmm: Add poweroff mode
- Add poweroff mode to VMM driver infrastructure and existing VMM
drivers.  When a driver is called in this mode, it must delete any
memory image that may exist for the given parcel.  In addition, if the
SUSPENDED flag matches the VMM driver being called, the driver must set
the flag to "no", even if there was no memory image to delete.