client: No longer support "resume -d" if not fully hoarded
- Previously, "isr resume -d" would perform a checkout and hoard, if
necessary, before resuming disconnected.  This is strange: if the user
asks to run disconnected, we shouldn't use a network connection, even
if the parcel is not prepared for disconnected operation.  Change
"resume -d" to fail if the parcel is not checked out and hoarded,
informing the user that they need to run "checkout -d" first.

- Update man page accordingly

- Update usage message, man page, and bash completions to remove
--checkstate support from resume mode, since the option was only
respected when disconnect() was run during resume, which we no longer do

- Don't call need_remote_parcel_and_lock() during resume if we don't
have a checkout, since we clearly won't have the lock yet