tools/disktool: Remove -n/--nchunks option
- Remove -n/--nchunks option, since it clutters up the code and doesn't
make a lot of sense: if the user wants to truncate an incoming/outgoing
image, they probably won't want to do it exactly on a chunk boundary,
and they may want to specify an initial offset as well.  The option is
only really useful when importing from /dev/zero, and we'll do that a
different way.

- Move read() call to loop test in import_image()