client: Don't allow "isr resume" to time out waiting for PK to exit
- At suspend time, the client would wait 60 seconds for PK to exit and
then give up.  Long suspend delays usually mean that the VMM has not
in fact suspended and is still using the virtual disk.  Therefore, the
timeout gives incorrect feedback to the user, who may unsafely proceed
with other operations such as checking in the parcel.  Instead, wait
indefinitely.  This also prevents the pm-utils hook from believing that
parcels have suspended successfully when they have not.

- No longer show the PK pid in the "Shutting down" message, since we
don't really want the user to kill PK themselves

- Use a spinner rather than marching dots while waiting for PK to exit
so that the terminal doesn't fill with dots

- If PK hasn't exited in 30 seconds, print an explanatory warning to the
user and continue waiting