pk: Be less optimistic about repaired keyrings
- When we discover SQLite integrity errors in the keyring, only drop
and recreate indexes if splice is requested, and always set the damaged
flag.  There was an incident in testing in which further keyring
corruption occurred after an index rebuild, implying other integrity
problems that were not detected by integrity_check.  Therefore, we do
enough to let the user get their data out of the parcel, but don't
allow checkin of a repaired keyring.