vmm/virtualbox: Update version checking for 3.1
VirtualBox 3.1 no longer auto-upgrades all machine settings files to
the newest version; instead, it tries to leave them in their original
format until a change is made which requires a newer format.  The
settingsFormatVersion global attribute has therefore been removed.

- We previously checked the supported settings version and, if it
was unrecognized, warned the user that the installed VirtualBox version
might not be supported.   Move the check from fixup_upgrade_settings()
to a dedicated function, check the VirtualBox version instead of the
supported settings version, and no longer require the user to assent
before continuing.

- Remove "Auto-converting VM settings file during unregistration" warning,
since we can no longer easily test for this

- Add and use compat functions to check whether the settings file
version will be automatically upgraded or is too new for the current
VBox.  If running on VBox 3.1 or above, return our best guess.