vmm/virtualbox: Configure new parcels with custom CPUID
- When configuring a new parcel under VBox >= 3.1, override the guest
CPUID data to ensure that the guest sees a consistent set of CPU features
regardless of the processor we're resuming on.  Note that this only
affects CPUID calls from guest supervisor mode, not guest user mode.

- When configuring a new parcel under VBox >= 3.1, disable VBox's CPUID
checks, allowing a parcel to resume on a machine with a different CPU
than the one on which it was suspended