pk: Fill entire cache slot when writing into the local cache
- When writing a newly-obtained chunk into the local cache, write
chunksize bytes, not just the actual bytes of compressed data.  Nexus
already does this when it writes to the local cache, so the old behavior
causes the filesystem to allocate blocks for the first half of the chunk
slot when it is first read and for the rest of the slot when it is first
written.  These two allocations may not be contiguous, forcing additional
seeks at runtime and checkin.  This change should help ensure that chunk
slots are allocated a contiguous range of blocks on disk.