nexus: Call add_disk() from dedicated thread
- Attach the Nexus block device to the block layer from a dedicated
workqueue thread, rather than using the shared kernel workqueue.  This
operation can block for a long time if the chunks in the partition table
and at the end of the disk need to be fetched from the server.  If
add_disk() is run from the shared workqueue, this will block unrelated
tasks enqueued by other parts of the kernel, such as processing of TTY
receive buffers.