Allow VMM drivers to request that root be permitted to access the VFS
- pk: Add --allow-root option in run mode, which requests that the FUSE
filesystem be mounted with -oallow_root.  This option allows the root
user to access the virtual filesystem, as required by some VMMs, but
will cause fuse_mount() to fail if "user_allow_other" is not set in
/etc/fuse.conf.

- vmm: Add USES_ROOT output value in info mode, which the VMM driver
should set to "yes" if the VMM needs to access the VFS as root.

- client: If the VMM sets USES_ROOT=yes, check that /etc/fuse.conf
is configured appropriately and bail out with a helpful error message if
not.  Otherwise, pass --allow-root to PK's run mode.